# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * The Dolls - LGBTQ+ Friendly Fashion E-Commerce API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  originalPrice?: number | null;
  imageUrl: string;
  images: string[];
  category: string;
  tags: string[];
  sizes: string[];
  colors: string[];
  inStock: boolean;
  stockCount: number;
  rating: number;
  reviewCount: number;
  isFeatured: boolean;
  isNew: boolean;
  isSale: boolean;
  lgbtqFriendly: boolean;
  createdAt: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  imageUrl: string;
  productCount: number;
}

export interface CartItem {
  id: number;
  productId: number;
  product: Product;
  quantity: number;
  size?: string | null;
  color?: string | null;
}

export interface Cart {
  items: CartItem[];
  subtotal: number;
  total: number;
  itemCount: number;
}

export interface AddToCartInput {
  productId: number;
  quantity: number;
  size?: string | null;
  color?: string | null;
}

export interface UpdateCartItemInput {
  quantity: number;
}

export interface WishlistInput {
  productId: number;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  processing: "processing",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface Order {
  id: number;
  status: OrderStatus;
  items: CartItem[];
  total: number;
  shippingAddress: string;
  createdAt: string;
}

export interface CreateOrderInput {
  shippingAddress: string;
  email: string;
}

export interface LiveEvent {
  id: number;
  title: string;
  description: string;
  startsAt: string;
  endsAt: string;
  isLive: boolean;
  viewerCount: number;
  thumbnailUrl: string;
  hostName: string;
  featuredProducts: number[];
}

export type FeedMessageType =
  (typeof FeedMessageType)[keyof typeof FeedMessageType];

export const FeedMessageType = {
  chat: "chat",
  purchase: "purchase",
  join: "join",
  heart: "heart",
} as const;

export interface FeedMessage {
  id: number;
  username: string;
  message: string;
  type: FeedMessageType;
  createdAt: string;
}

export interface LiveEventDetail {
  event?: LiveEvent | null;
  products: Product[];
  feed: FeedMessage[];
}

export interface ShopSummary {
  totalProducts: number;
  newArrivals: number;
  onSale: number;
  categories: number;
  featuredProducts: Product[];
  newArrivalsProducts: Product[];
  saleProducts: Product[];
}

export type GetProductsParams = {
  category?: string;
  search?: string;
  sort?: GetProductsSort;
  limit?: number;
  offset?: number;
};

export type GetProductsSort =
  (typeof GetProductsSort)[keyof typeof GetProductsSort];

export const GetProductsSort = {
  newest: "newest",
  price_asc: "price_asc",
  price_desc: "price_desc",
  popular: "popular",
} as const;

export type AddToWishlist200 = {
  success: boolean;
};

export type RemoveFromWishlist200 = {
  success: boolean;
};