# lib/api-zod/src/generated/types/liveEvent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * The Dolls - LGBTQ+ Friendly Fashion E-Commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface LiveEvent {
  id: number;
  title: string;
  description: string;
  startsAt: string;
  endsAt: string;
  isLive: boolean;
  viewerCount: number;
  thumbnailUrl: string;
  hostName: string;
  featuredProducts: number[];
}