# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * The Dolls - LGBTQ+ Friendly Fashion E-Commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  originalPrice?: number | null;
  imageUrl: string;
  images: string[];
  category: string;
  tags: string[];
  sizes: string[];
  colors: string[];
  inStock: boolean;
  stockCount: number;
  rating: number;
  reviewCount: number;
  isFeatured: boolean;
  isNew: boolean;
  isSale: boolean;
  lgbtqFriendly: boolean;
  createdAt: string;
}