# lib/api-zod/src/generated/types/shopSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * The Dolls - LGBTQ+ Friendly Fashion E-Commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { Product } from "./product";

export interface ShopSummary {
  totalProducts: number;
  newArrivals: number;
  onSale: number;
  categories: number;
  featuredProducts: Product[];
  newArrivalsProducts: Product[];
  saleProducts: Product[];
}