# lib/api-zod/src/generated/types/wishlistInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * The Dolls - LGBTQ+ Friendly Fashion E-Commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface WishlistInput {
  productId: number;
}